pyproject.toml
[project]
name = "skypilot-mcp"
version = "0.1.0"
description = "MCP server for SkyPilot - manage cloud clusters, jobs, and resources via Model Context Protocol"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=3.0.0",
    "skypilot>=0.11.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pre-commit>=4.0",
    "ruff>=0.3",
]

[project.scripts]
skypilot-mcp = "skypilot_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skypilot_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
